AI_Coding_Agent/pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai,tools]>=1.7.0",
    "python-dotenv>=1.1.1",
]
